+++ NEWS.md
@@ -3,6 +3,8 @@
 * Added `ResamplingLOO` for leave-one-out resampling.
 * Fixed `ResamplingBootstrap` in combination with grouping (#514).
 * Fixed plot method of `TaskGeneratorMoons`.
+* Added hyperparameter `keep_model` to learners `"classif.rpart"` and
+  `"regr.rpart"`.
 
 # mlr3 0.4.0